+++ hub/experimental/__init__.py
@@ -1,2 +1,2 @@
 from hub.experimental.dataloader import dataloader
-from hub.experimental.query import query
+from hub.experimental.hub3_query import query
